refactor(FilterTypeFileTypes): extract selection state to useContentTypeSelection composable
